[REFACTOR] Terminology download=>download_cache
This PR renames download to download_cache for better clarity.
